:truck: rename package "dto" to "model"
