+++ src/main/java/cc/zoyn/core/model/package-info.java
@@ -3,4 +3,4 @@
  * @author Zoyn
  * @since 2017-12-09
  */
-package cc.zoyn.core.dto;
+package cc.zoyn.core.model;
